execute the build in a separate container
The controller now inspects the "language" key in the config
and spawns the corresponding container. It then sends the shell
commands from each config section to the container and commits after
each step. When a step fails, it is possible to inspect the state of
any prior step.

For the moment, the controller must be executed inside a container
itself because the volume mount for the repo is hardcoded to be
from the controller container.